requirements.txt
streamlit>=1.35
pandas>=2.0
plotly>=5.18
requests>=2.31
python-dotenv>=1.0
